+++ staging/src/components/Dropdown/index.ts
@@ -1,1 +1,13 @@
+import ToggleRight from "./ToggleRight.astro";
+import ToggleLeft from "./ToggleLeft.astro";
+import BelowUpwards from "./BelowUpwards.astro";
+import ToggleRightUpwards from "./ToggleRightUpwards.astro";
+import ToggleLeftUpwards from "./ToggleLeftUpwards.astro";
 export * from "./Dropdown";
+export {
+  ToggleRight,
+  ToggleLeft,
+  BelowUpwards,
+  ToggleRightUpwards,
+  ToggleLeftUpwards,
+};